BonusApp Core,Data,Services

+++ Libraries/Nop.Core/Domain/BonusApp/Customers/Customer.cs
@@ -77,7 +77,7 @@ namespace Nop.Core.Domain.BonusApp.Customers
 
         #region Bonus properties
                   
-        public double Money { get; set; }
+        public decimal Money { get; set; }
 
         #endregion
